+++ packages/geo/src/lib/draw/draw/draw.component.ts
@@ -23,21 +23,24 @@ import {
 
 import { LanguageService } from '@igo2/core';
 import { MatDialog } from '@angular/material/dialog';
-import { FontType, GeometryType } from '../shared/draw.enum';
+import { CoordinatesUnit, FontType, GeometryType, LabelType } from '../shared/draw.enum';
 import { IgoMap } from '../../map/shared/map';
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { Draw, FeatureWithDraw } from '../shared/draw.interface';
-import { UntypedFormGroup, UntypedFormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, UntypedFormControl } from '@angular/forms';
 import { VectorSourceEvent as OlVectorSourceEvent } from 'ol/source/Vector';
 import { VectorLayer } from '../../layer/shared/layers/vector-layer';
 import { FeatureDataSource } from '../../datasource/shared/datasources/feature-datasource';
 import { DrawControl } from '../../geometry/shared/controls/draw';
 import {
   EntityRecord,
+  EntityTableButton,
+  EntityTableColumnRenderer,
   EntityTableTemplate
 } from '@igo2/common';
 
 import * as OlStyle from 'ol/style';
+import * as olproj from 'ol/proj';
 import OlVectorSource from 'ol/source/Vector';
 import OlCircle from 'ol/geom/Circle';
 import OlPoint from 'ol/geom/Point';
@@ -46,13 +49,13 @@ import OlGeoJSON from 'ol/format/GeoJSON';
 import OlOverlay from 'ol/Overlay';
 import type { Type } from 'ol/geom/Geometry';
 import { default as OlGeometry } from 'ol/geom/Geometry';
-import { getDistance } from 'ol/sphere';
+import { getDistance, getLength } from 'ol/sphere';
 import { DrawStyleService } from '../shared/draw-style.service';
 import { first, skip } from 'rxjs/operators';
 import { DrawPopupComponent } from './draw-popup.component';
 import { DrawShorcutsComponent } from './draw-shorcuts.component';
 import { getTooltipsOfOlGeometry } from '../../measure/shared/measure.utils';
-import { createInteractionStyle } from '../shared/draw.utils';
+import { createInteractionStyle, DDtoDMS } from '../shared/draw.utils';
 import { transform } from 'ol/proj';
 import { DrawIconService } from '../shared/draw-icon.service';
 import { StyleModalComponent, StyleModalData } from '../../layer/style-modal/style-modal.component';
@@ -64,9 +67,24 @@ import {
   animate,
   transition
 } from '@angular/animations';
-import Point from 'ol/geom/Point';
 
 import { DrawLayerPopupComponent } from './draw-layer-popup.component';
+
+import {
+  measureOlGeometryLength,
+  measureOlGeometryArea,
+  metersToUnit,
+  squareMetersToUnit
+} from '../../measure/shared/measure.utils';
+
+import {
+  MeasureLengthUnit,
+  MeasureLengthUnitAbbreviation,
+  MeasureAreaUnit,
+  MeasureAreaUnitAbbreviation,
+} from '../../measure/shared/measure.enum';
+import Polygon, { fromCircle } from 'ol/geom/Polygon';
+
 
 @Component({
   selector: 'igo-draw',
@@ -108,10 +126,24 @@ export class DrawComponent implements OnInit, OnDestroy {
       {
         name: 'Drawing',
         title: this.languageService.translate.instant('igo.geo.draw.labels'),
-        tooltip: this.languageService.translate.instant('igo.geo.draw.changeLabel'),
         valueAccessor: (feature: FeatureWithDraw) => {
           return feature.properties.draw;
         }
+      },
+      {
+        name: 'Edition',
+        title: '',
+        sort: false,
+        valueAccessor: (feature: FeatureWithDraw) => {
+          return [{
+            editMode: false,
+            icon: 'pencil',
+            color: 'primary',
+            click: () => { this.editLabelDrawing(feature);},
+            style: 'mat-icon-button'
+          }] as EntityTableButton[];
+        },
+        renderer: EntityTableColumnRenderer.ButtonGroup,
       }
     ]
   };
@@ -171,6 +203,7 @@ export class DrawComponent implements OnInit, OnDestroy {
   public drawControlIsDisabled: boolean = true;
   public drawControlIsActive: boolean = false;
   public labelsAreShown: boolean;
+  public freehandMode = false;
   private subscriptions$$: Subscription[] = [];
 
   public position: string = 'bottom';
@@ -178,6 +211,20 @@ export class DrawComponent implements OnInit, OnDestroy {
   public icons: Array<string>;
   public icon: string;
 
+  public radiusFormControl = new UntypedFormControl(1000);
+  public measureUnit: MeasureLengthUnit = MeasureLengthUnit.Meters;
+  public radiusFormControlChange$$: Subscription = new Subscription();
+  public predefinedRadius$: BehaviorSubject<number> = new BehaviorSubject(undefined);
+  public radiusDrawEnd$: BehaviorSubject<number> = new BehaviorSubject(undefined);
+
+
+  /**
+   * Available measure units for the measure type given
+   * @internal
+   */
+  get measureUnits(): string[] {
+    return [MeasureLengthUnit.Meters, MeasureLengthUnit.Kilometers];
+  }
   private numberOfDrawings: number;
   public isCreatingNewLayer: boolean = false;
   private currGeometryType = this.geometryType.Point as any;
@@ -190,6 +237,7 @@ export class DrawComponent implements OnInit, OnDestroy {
     private drawStyleService: DrawStyleService,
     private dialog: MatDialog,
     private drawIconService: DrawIconService
+
   ) {
     this.buildForm();
     this.fillColor = this.drawStyleService.getFillColor();
@@ -270,6 +318,12 @@ export class DrawComponent implements OnInit, OnDestroy {
       });
       this.onLayerChange(this.activeDrawingLayer);
     }
+
+    this.radiusFormControlChange$$ = this.radiusFormControl.valueChanges.subscribe(value => {
+      if (this.activeDrawControl.ispredefinedRadius$.getValue()) {
+        this.changeRadius(value);
+      }
+    });
   }
 
   /**
@@ -281,6 +335,7 @@ export class DrawComponent implements OnInit, OnDestroy {
     this.activeStore.state.updateAll({selected: false});
     this.deactivateDrawControl();
     this.subscriptions$$.map((s) => s.unsubscribe());
+    this.radiusFormControlChange$$.unsubscribe();
   }
 
   /**
@@ -374,39 +429,34 @@ export class DrawComponent implements OnInit, OnDestroy {
       // open the dialog box used to enter label
       const dialogRef = this.dialog.open(DrawPopupComponent, {
         disableClose: false,
-        data: { currentLabel: olGeometry.get('draw') }
+        data: {olGeometry: olGeometry, map: this.map}
       });
 
       // when dialog box is closed, get label and set it to geometry
-      dialogRef.afterClosed().subscribe((label: string) => {
+      dialogRef.afterClosed().subscribe((result) => {
         // checks if the user clicked ok
         if (dialogRef.componentInstance.confirmFlag) {
-          this.updateLabelOfOlGeometry(olGeometry, label);
-          if (!olGeometry.values_.fontStyle) {
-            this.updateFontSizeAndStyle(olGeometry, '20', FontType.Arial);
-          }
-          if (!olGeometry.values_.drawingStyle) {
+
+          this.updateLabelOfOlGeometry(olGeometry, result.label);
+          this.updateLabelType(olGeometry, dialogRef.componentInstance.labelFlag);
+          this.updateMeasureUnit(olGeometry, result.measureUnit);
+
+          if (!(olGeometry instanceof OlFeature)){
+            this.updateFontSizeAndStyle(olGeometry, '15', FontType.Arial);
             this.updateFillAndStrokeColor(
               olGeometry,
               'rgba(255,255,255,0.4)',
               'rgba(143,7,7,1)'
             );
-          }
-          if (!(olGeometry.values_.offsetX || olGeometry.values_.offsetY)) {
             this.updateOffset(
               olGeometry,
               0,
-              olGeometry instanceof Point ? -15 : 0
+              (olGeometry instanceof OlPoint) ? -15 : 0
             );
           }
 
-          // if event was fired at draw end
-          if (isDrawEnd) {
-            this.onDrawEnd(olGeometry);
-            // if event was fired at select
-          } else {
-            this.onSelectDraw(olGeometry, label);
-          }
+          isDrawEnd ? this.onDrawEnd(olGeometry): this.onSelectDraw(olGeometry, result.label,
+            [dialogRef.componentInstance.labelFlag, result.measureUnit]);
           this.updateHeightTable();
         }
         // deletes the feature
@@ -441,11 +491,63 @@ export class DrawComponent implements OnInit, OnDestroy {
 
     entities.forEach((entity) => {
       const entityId = entity.properties.id;
-
       const olGeometryId = olGeometry.ol_uid;
 
       if (entityId === olGeometryId) {
-        this.updateLabelOfOlGeometry(olGeometry, entity.properties.draw);
+        if (entity.properties.labelType === LabelType.Coordinates){
+          let longLat = DDtoDMS([entity.properties.longitude, entity.properties.latitude],
+            entity.properties.measureUnit as CoordinatesUnit);
+          this.updateLabelOfOlGeometry(olGeometry, '(' + longLat[1] + ', ' + longLat[0] + ')');
+        }
+        else if (entity.properties.labelType === LabelType.Length){
+          if (olGeometry instanceof OlCircle){
+            let circularPolygon = fromCircle(olGeometry, 10000);
+            const radius = metersToUnit(this.getRadius(circularPolygon), entity.properties.measureUnit as MeasureLengthUnit);
+            const unit = MeasureLengthUnitAbbreviation[entity.properties.measureUnit];
+            const radiusLabel = 'R: ' + radius.toFixed(2).toString() + ' ' + unit;
+            this.updateLabelOfOlGeometry(olGeometry, radiusLabel);
+          }
+          else {
+            let olGeometryLength = measureOlGeometryLength(olGeometry, this.map.ol.getView().getProjection().getCode());
+            let measureUnit: any;
+            const temp: MeasureLengthUnit = entity.properties.measureUnit as MeasureLengthUnit;
+            measureUnit = MeasureLengthUnitAbbreviation[entity.properties.measureUnit];
+            olGeometryLength = metersToUnit(olGeometryLength, temp);
+            let lengthLabel = olGeometry instanceof Polygon ?
+              'P: ' + olGeometryLength.toFixed(2).toString() + ' ' + measureUnit
+              : olGeometryLength.toFixed(2).toString() + ' ' + measureUnit;
+            this.updateLabelOfOlGeometry(olGeometry, lengthLabel);
+          }
+        }
+        else if (entity.properties.labelType === LabelType.Area){
+          if (olGeometry instanceof OlCircle){
+            let circularPolygon = fromCircle(olGeometry, 10000);
+            let circleArea = measureOlGeometryArea(circularPolygon, this.map.ol.getView().getProjection().getCode());
+            const unit = MeasureAreaUnitAbbreviation[entity.properties.measureUnit];
+            const temp: MeasureAreaUnit = entity.properties.measureUnit as MeasureAreaUnit;
+            circleArea = squareMetersToUnit(circleArea, temp);
+            const areaLabel = circleArea.toFixed(2).toString() + ' ' + unit;
+            this.updateLabelOfOlGeometry(olGeometry, areaLabel);
+          }
+          else {
+            let olGeometryArea = measureOlGeometryArea(olGeometry, this.map.ol.getView().getProjection().getCode());
+            let measureUnit: any;
+            const temp: MeasureAreaUnit = entity.properties.measureUnit as MeasureAreaUnit;
+            measureUnit = MeasureAreaUnitAbbreviation[entity.properties.measureUnit];
+            olGeometryArea = squareMetersToUnit(olGeometryArea, temp);
+            const lengthLabel = olGeometryArea.toFixed(2).toString() + ' ' + measureUnit;
+            this.updateLabelOfOlGeometry(olGeometry, lengthLabel);
+          }
+        }
+        else {
+          this.updateLabelOfOlGeometry(olGeometry, entity.properties.draw);
+        }
+
+        this.updateLabelType(
+          olGeometry,
+          entity.properties.labelType
+        );
+        this.updateMeasureUnit(olGeometry, entity.properties.measureUnit);
         this.updateFontSizeAndStyle(
           olGeometry,
           entity.properties.fontStyle.split(' ')[0].replace('px', ''),
@@ -468,7 +570,7 @@ export class DrawComponent implements OnInit, OnDestroy {
     });
   }
 
-  private onSelectDraw(olFeature: OlFeature<OlGeometry>, label: string) {
+  private onSelectDraw(olFeature: OlFeature<OlGeometry>, label: string, labelTypeAndUnit?) {
     const entities = this.activeStore.all();
 
     const olGeometry = olFeature.getGeometry() as any;
@@ -502,6 +604,8 @@ export class DrawComponent implements OnInit, OnDestroy {
         this.updateFontSizeAndStyle(olGeometry, fontSize, fontStyle);
         this.updateFillAndStrokeColor(olGeometry, fillColor, strokeColor);
         this.updateOffset(olGeometry, offsetX, offsetY);
+        this.updateLabelType(olGeometry, labelTypeAndUnit[0]);
+        this.updateMeasureUnit(olGeometry, labelTypeAndUnit[1]);
         this.replaceFeatureInStore(entity, olGeometry, rad);
       }
     });
@@ -515,13 +619,14 @@ export class DrawComponent implements OnInit, OnDestroy {
   private addFeatureToStore(
     olGeometry,
     radius?: number,
-    feature?: FeatureWithDraw
+    feature?: FeatureWithDraw,
   ) {
     let rad: number;
     let center4326: Array<number>;
     let point4326: Array<number>;
     let lon4326: number;
     let lat4326: number;
+
     const featureId = feature ? feature.properties.id : olGeometry.ol_uid;
     const projection = this.map.ol.getView().getProjection();
 
@@ -555,7 +660,12 @@ export class DrawComponent implements OnInit, OnDestroy {
         lon4326 = center4326[0];
         lat4326 = center4326[1];
         rad = getDistance(center4326, extent4326);
+        this.radiusFormControl.setValue(Math.round(rad));
       }
+
+    }
+    if (this.activeDrawControl.radiusDrawEnd$.getValue()) {
+      rad = this.activeDrawControl.radiusDrawEnd$.getValue();
     }
 
     if (olGeometry instanceof OlPoint) {
@@ -584,12 +694,16 @@ export class DrawComponent implements OnInit, OnDestroy {
           stroke: olGeometry.get('strokeColor_')
         },
         offsetX: olGeometry.get('offsetX_'),
-        offsetY: olGeometry.get('offsetY_')
+        offsetY: olGeometry.get('offsetY_'),
+        labelType: olGeometry.get('labelType_'),
+        measureUnit: olGeometry.get('measureUnit_')
       },
       meta: {
         id: featureId
       }
     });
+    this.activeDrawControl.predefinedRadius$.next(undefined);
+    this.activeDrawControl.radiusDrawEnd$.next(undefined);
   }
 
   private buildForm() {
@@ -641,14 +755,13 @@ export class DrawComponent implements OnInit, OnDestroy {
   /**
    * Called when the user double-clicks the selected drawing
    */
-  editLabelDrawing() {
-    if (this.selectedFeatures$.value.length) {
-      const olGeometry = featureToOl(
-        this.selectedFeatures$.value[0],
-        this.map.ol.getView().getProjection().getCode()
-      );
-      this.openDrawDialog(olGeometry, false);
-    }
+  editLabelDrawing(feature) {
+    const olGeometryFeature = featureToOl(
+      feature,
+      this.map.ol.getView().getProjection().getCode()
+    );
+    this.openDrawDialog(olGeometryFeature, false);
+
   }
 
   openShorcutsDialog() {
@@ -694,6 +807,19 @@ export class DrawComponent implements OnInit, OnDestroy {
    */
   onToggleDrawControl(toggleIsChecked: boolean) {
     toggleIsChecked ? this.toggleDrawControl() : this.deactivateDrawControl();
+  }
+
+  onToggleFreehandMode(event: any) {
+    if (this.isCircle() && !event.checked) {
+      this.activeDrawControl.ispredefinedRadius$.next(true);
+      this.changeRadius(this.radiusFormControl.value);
+    } else {
+      this.activeDrawControl.setOlInteractionStyle(createInteractionStyle(this.fillColor, this.strokeColor, this.strokeWidth));
+      this.activeDrawControl.ispredefinedRadius$.next(false);
+    }
+    this.freehandMode = event.checked;
+    this.activeDrawControl.freehand$.next(event.checked);
+    this.toggleDrawControl();
   }
 
   // User changes properties of a drawing element
@@ -900,6 +1026,9 @@ export class DrawComponent implements OnInit, OnDestroy {
   onGeometryTypeChange(geometryType: Type) {
     this.currGeometryType = geometryType;
     this.activeDrawControl.setGeometryType(geometryType);
+    this.freehandMode ?
+      this.onToggleFreehandMode({ checked: true }) :
+      this.onToggleFreehandMode({ checked: false });
     this.toggleDrawControl();
   }
 
@@ -907,11 +1036,11 @@ export class DrawComponent implements OnInit, OnDestroy {
 
   /**
    * Update the label of a geometry when a label is entered in a dialog box
-   * @param olGeometry the geometry
+   * @param OlFeature the feature
    * @param label the label
    */
-  private updateLabelOfOlGeometry(olGeometry: OlGeometry, label: string) {
-    olGeometry.setProperties(
+  private updateLabelOfOlGeometry(OlFeature, label: string) {
+    OlFeature.setProperties(
       {
         _label: label
       },
@@ -960,6 +1089,31 @@ export class DrawComponent implements OnInit, OnDestroy {
     );
   }
 
+  private updateLabelType(
+    olFeature: OlFeature<OlGeometry>,
+    typeOfLabel: LabelType | [LabelType, LabelType]
+  ){
+    olFeature.setProperties(
+      {
+        labelType_:typeOfLabel
+      },
+      true
+    );
+  }
+
+  private updateMeasureUnit(
+    olFeature: OlFeature<OlGeometry>,
+    measureUnit: MeasureLengthUnit | MeasureAreaUnit | CoordinatesUnit | []
+  ){
+    olFeature.setProperties(
+      {
+        measureUnit_: measureUnit
+      },
+      true
+    );
+
+  }
+
   // Updates values of the selected element on the HTML view
 
   getFeatureFontSize(): string {
@@ -967,7 +1121,7 @@ export class DrawComponent implements OnInit, OnDestroy {
       ? this.selectedFeatures$.value[0].properties.fontStyle
           .split(' ')[0]
           .replace('px', '')
-      : '20';
+      : '15';
   }
 
   getFeatureFontStyle() {
@@ -1102,8 +1256,77 @@ export class DrawComponent implements OnInit, OnDestroy {
         }
       );
     }
-
     this.activeDrawControl.setOlMap(this.map.ol, true);
+  }
+
+  /**
+   * get the geometry of design
+   *
+   */
+
+  isPoint() {
+    return this.activeDrawControl.getGeometryType() === this.geometryType.Point;
+  }
+
+  isLineString() {
+    return this.activeDrawControl.getGeometryType() === this.geometryType.LineString;
+  }
+
+  isPolygon() {
+    return this.activeDrawControl.getGeometryType() === this.geometryType.Polygon;
+  }
+
+  isCircle() {
+    return this.activeDrawControl.getGeometryType() === this.geometryType.Circle;
+  }
+
+  /**
+   * The fonction to predefine the radius of the user
+   *
+   */
+
+  changeRadius(radius: number) {
+    let radiusMeters: number;
+
+    if (radius) {
+      this.measureUnit === MeasureLengthUnit.Meters ? radiusMeters = radius :
+        radiusMeters = radius * 1000;
+    } else {
+      radiusMeters = undefined;
+    }
+
+    const pointStyle = (feature: OlFeature<OlGeometry>, resolution: number) => {
+      const geom = feature.getGeometry() as OlPoint;
+      const coordinates = olproj.transform(geom.getCoordinates(), this.map.projection, 'EPSG:4326');
+
+      const radius = radiusMeters / (Math.cos((Math.PI / 180) * coordinates[1])) / resolution;
+      this.activeDrawControl.predefinedRadius$.next(radiusMeters);
+      return new OlStyle.Style({
+        image: new OlStyle.Circle({
+          radius: radius,
+          stroke: new OlStyle.Stroke({
+            width: 1,
+            color: 'rgba(143,7,7,1)'
+          }),
+          fill: new OlStyle.Fill({
+            color: 'rgba(255,255,255,0.4)'
+          })
+        })
+      });
+    };
+
+    this.activeDrawControl.setOlInteractionStyle(pointStyle);
+    this.toggleDrawControl();
+  }
+
+  onMeasureUnitChange(selectedMeasureUnit: MeasureLengthUnit) {
+    if (selectedMeasureUnit === this.measureUnit) {
+      return;
+    } else {
+      this.measureUnit = selectedMeasureUnit;
+      this.measureUnit === MeasureLengthUnit.Meters ? this.radiusFormControl.setValue(this.radiusFormControl.value * 1000) :
+        this.radiusFormControl.setValue(this.radiusFormControl.value / 1000);
+    }
   }
 
   /**
@@ -1139,4 +1362,10 @@ export class DrawComponent implements OnInit, OnDestroy {
       });
     }
   }
+
+  private getRadius(olGeometry): number{
+    const length = getLength(olGeometry);
+    return Number(length / (2 * Math.PI));
+  }
+
 }