+++ packages/geo/src/lib/import-export/shared/import.interface.ts
@@ -1,7 +1,9 @@
+import { ExportFormat } from "./export.type";
+
 export interface ImportExportServiceOptions {
   url: string;
   clientSideFileSizeMaxMb?: number;
   forceNaming?: boolean;
-  formats?: string[];
+  formats?: ExportFormat[];
   configFileToGeoDBService?: string;
 }